# remove branch vs leaf node restriction in index