@woodsaj 2017-01-20
This is mostly done, just need to add unit tests to verify that the index works for mixed branch/leaf nodes