@woodsaj 2017-01-20
no significant change in performance of the index.

old
```
anthony:~/go/src/github.com/raintank/metrictank/idx/memory$ go test -run NONE -bench=. -benchmem 
BenchmarkFind-4              	   10000	    184041 ns/op	   51589 B/op	     877 allocs/op
BenchmarkConcurrent4Find-4   	   20000	    151151 ns/op	   51600 B/op	     877 allocs/op
BenchmarkConcurrent8Find-4   	   20000	    154791 ns/op	   51576 B/op	     877 allocs/op
BenchmarkIndexing-4          	  500000	      3658 ns/op	    1155 B/op	      22 allocs/op
PASS

``

new 
```
anthony:~/go/src/github.com/raintank/metrictank/idx/memory$ go test -run NONE -bench=. -benchmem 
BenchmarkFind-4              	   10000	    196579 ns/op	   53480 B/op	     877 allocs/op
BenchmarkConcurrent4Find-4   	   20000	    183117 ns/op	   53487 B/op	     877 allocs/op
BenchmarkConcurrent8Find-4   	   20000	    155884 ns/op	   53451 B/op	     876 allocs/op
BenchmarkIndexing-4          	  500000	      3754 ns/op	    1187 B/op	      22 allocs/op
PASS

```