@woodsaj 2017-01-30
lets get PR #504 merged, then i will tackle this again.