@woodsaj 2017-02-02
This is now ready for review.

Unfortunately, the Grafana query editor is unable to view series that are a descendant of a leaf node as graphite still does not correctly support this.

When you have a series called "foo.bar" and a second series called "foo.bar.baz", carbon will write a whisper file for each.  However the graphite "/metrics/find?" query only supports sending back that a node is a leaf or a branch and not both due to:

https://github.com/graphite-project/graphite-web/blob/master/webapp/graphite/metrics/views.py#L270-L283

* graphite-api does the exact same thing

However, if you set the Grafana queryEditor to raw mode, you can add the full series path and it will render.


When we add graphite query handling directly to MT, we can fix this limitation.